+++ CHANGELOG.md
@@ -5,6 +5,7 @@ We follow [Semantic Versions](https://semver.org/).
 # Version 2.0.0
 
 - Use updated `qaseio` sdk (^4.0.0)
+- Rework getting browser name (see `README.md`)
 - Drop `Python 3.10` support
 - Drop `pytest-variables` support (now use pytest options to provide browser name)
 - Add `ruff` and `cspell` linters